[verified] fix(plugins): address marketplace review feedback
